prod filter hook setup
